Move `@RequiresDefaultInjection` to API
