+++ BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/api/core/service/annotations/RequiresDefaultInjection.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.botcommands.internal.core.service.annotations
+package io.github.freya022.botcommands.api.core.service.annotations
 
 import io.github.freya022.botcommands.internal.core.service.BCInjectionCondition
 import org.springframework.context.annotation.Conditional
@@ -7,4 +7,4 @@ import org.springframework.context.annotation.Conditional
  * Makes a service disabled when using Spring
  */
 @Conditional(BCInjectionCondition::class)
-internal annotation class RequiresDefaultInjection
+annotation class RequiresDefaultInjection
